Hosting some conversations about performance in this PR. Still trying to figure out. With the script in `inst/bench/bench.R` here some initial results: 

```r
benchs(
  iterations = 10,
  libs = libs,
  setup = {
    df <- tibble(x = rnorm(1e6), y = rnorm(1e6), g = sample(rep(1:1e4, 100))) %>% group_by(g)
  },
  summarise(df, n = 1 + 1),
  summarise(df, n = +n()),

  summarise(df, n = length(x)),
  summarise(df, n = +mean(x)),
  summarise(df, n = mean.default(x)),
  summarise(df, n = .Internal(mean(x))),

  summarise(df, nx = length(x), ny = length(y)),
  summarise(df, mean(x) + mean(y))
)
```

```
# A tibble: 8 x 3
  expression                                     `0.8.3`   master
  <chr>                                         <bch:tm> <bch:tm>
1 summarise(df, n = 1 + 1)                        17.4ms   7.75ms
2 summarise(df, n = +n())                         31.4ms  87.46ms
3 summarise(df, n = length(x))                    35.1ms  25.74ms
4 summarise(df, n = +mean(x))                     66.9ms  63.49ms
5 summarise(df, n = mean.default(x))              48.8ms  36.89ms
6 summarise(df, n = .Internal(mean(x)))           34.9ms  27.06ms
7 summarise(df, nx = length(x), ny = length(y))   75.2ms  52.77ms
8 summarise(df, mean(x) + mean(y))               124.7ms 107.67ms
```

This is encouraging, to see at least non hybrid operations with similar or better performance. not the same story with fewer groups (rows = 1e6, n() = 10): 

```
# A tibble: 8 x 3
  expression                                     `0.8.3`   master
  <chr>                                         <bch:tm> <bch:tm>
1 summarise(df, n = 1 + 1)                       146.7µs    1.5ms
2 summarise(df, n = +n())                       167.64µs   1.56ms
3 summarise(df, n = length(x))                    7.48ms   9.81ms
4 summarise(df, n = +mean(x))                     10.9ms   10.7ms
5 summarise(df, n = mean.default(x))              9.25ms  11.99ms
6 summarise(df, n = .Internal(mean(x)))           10.8ms  13.07ms
7 summarise(df, nx = length(x), ny = length(y))  17.16ms   19.5ms
8 summarise(df, mean(x) + mean(y))               21.34ms  24.44ms
```

Will investigate. 

One thing that is becoming visible is that, on simple operations the timing is dominated by `vec_chop()`, which is already pretty optimized, but perhaps this is a place that could benefit from parallelization, at least for simple types.