@romainfrancois 2020-02-11
with proof of concept parallel `vec_chop()`: 

- nrows = 1e6, 10'000 groups of 100

```
# A tibble: 8 x 3
  expression                                     `0.8.3`   master
  <chr>                                         <bch:tm> <bch:tm>
1 summarise(df, n = 1 + 1)                        18.7ms   7.35ms
2 summarise(df, n = +n())                         31.1ms  80.17ms
3 summarise(df, n = length(x))                    35.6ms  16.45ms
4 summarise(df, n = +mean(x))                     67.5ms  46.12ms
5 summarise(df, n = mean.default(x))              49.3ms  29.07ms
6 summarise(df, n = .Internal(mean(x)))           36.4ms   18.5ms
7 summarise(df, nx = length(x), ny = length(y))   76.1ms  32.18ms
8 summarise(df, mean(x) + mean(y))               122.8ms  79.23ms
```

- nrows = 1e6, 10 groups of 100'000

```
# A tibble: 8 x 3
  expression                                     `0.8.3`   master
  <chr>                                         <bch:tm> <bch:tm>
1 summarise(df, n = 1 + 1)                       146.4µs   1.56ms
2 summarise(df, n = +n())                       165.05µs   1.64ms
3 summarise(df, n = length(x))                    8.89ms   3.13ms
4 summarise(df, n = +mean(x))                    10.89ms   5.47ms
5 summarise(df, n = mean.default(x))              9.45ms   5.49ms
6 summarise(df, n = .Internal(mean(x)))          11.12ms   5.28ms
7 summarise(df, nx = length(x), ny = length(y))  14.34ms   4.83ms
8 summarise(df, mean(x) + mean(y))               19.77ms   9.52ms
```